# api_docs.yaml
openapi: 3.0.0
info:
  title: LaunchPad Placement Portal API
  version: 1.0.0
  description: |
    REST API for the LaunchPad campus placement portal.

    **Authentication:** Most endpoints require an `Authentication-Token` header.
    Log in via `/auth/login` to get your token, then click "Authorize" above and paste it.
servers:
  - url: http://localhost:5001/api
paths:
...
  /admin/companies:
    get:
      summary: List companies
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [approved, pending, blacklisted] }
          description: Filter companies by status
        - name: search
          in: query
          schema: { type: string }
          description: Search by name or industry
      responses:
        "200": { description: Companies list }
        "403": { description: Forbidden — admin role required }
  /admin/companies/{company_id}/approve:
    put:
      summary: Approve company
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: company_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Company approved }
  /admin/companies/{company_id}/reject:
    put:
      summary: Reject company
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: company_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Company rejected }
  /admin/companies/{company_id}/blacklist:
    put:
      summary: Toggle company blacklist
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: company_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Blacklist status toggled }
  /admin/companies/{company_id}/deactivate:
    put:
      summary: Toggle company active status
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: company_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Active status toggled }
  /admin/drives:
    get:
      summary: List placement drives
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, closed] }
          description: Filter drives by status
      responses:
        "200": { description: Drives list }
  /admin/drives/{drive_id}/approve:
    put:
      summary: Approve drive
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Drive approved }
  /admin/drives/{drive_id}/reject:
    put:
      summary: Reject drive
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Drive rejected }
  /admin/students:
    get:
      summary: List students
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Search by name, phone, or branch
      responses:
        "200": { description: Students list }
  /admin/students/{student_id}/blacklist:
    put:
      summary: Toggle student blacklist
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: student_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Blacklist status toggled }
  /admin/students/{student_id}/deactivate:
    put:
      summary: Toggle student active status
      tags: [Admin]
      security:
        - TokenAuth: []
      parameters:
        - { name: student_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Active status toggled }
  /admin/analytics:
    get:
      summary: Admin dashboard analytics
      description: Cached for 120 seconds.
      tags: [Admin]
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Analytics data with summary, status breakdown, top companies, branch stats

  # Company
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authentication-Token
      description: Paste the token returned by POST /auth/login
...